Refactor: 2번 삭제 버그 수정과 삭제 로직 리팩토링
