+++ src/routes/myPlantRoutes.tsx
@@ -12,16 +12,12 @@ const MyPlantRegisterPage = lazy(
 const MyPlantEditPage = lazy(
   () => import('@/pages/myPlantPage/myPlantEditPage/MyPlantEditPage'),
 );
-const MyPlantRegisterPage2 = lazy(
-  () => import('@/pages/myPlantPage/myPlantRegisterPage/MyPlantRegisterPage2'),
-);
 
 const myPlantRoutes = [
   { path: '/myplant', element: <MyPlantMainPage /> },
   { path: '/myplant/:docId', element: <MyPlantDetailPage /> },
   { path: '/myplant/:docId/edit', element: <MyPlantEditPage /> },
   { path: '/myplant/register', element: <MyPlantRegisterPage /> },
-  { path: '/myplant/register2', element: <MyPlantRegisterPage2 /> },
 ];
 
 export default myPlantRoutes;
